solving issues with memory eeprom now it works with ram from ds1307

+++ Model.h
@@ -25,10 +25,10 @@ public:
 	uint32_t whenItStarted;
 	uint32_t whenToturnOff;
 
-	int hourPG1;
-	int minutePG1;
-	int minutesOnPG1;
-	int dayPG1;
+	uint8_t hourPG1;
+	uint8_t minutePG1;
+	uint8_t minutesOnPG1;
+	uint8_t dayPG1;
 
 	void savePG1();
